+++ YouboraLib/adapter/YBPlaybackFlags.h
@@ -41,9 +41,17 @@
 
 /// Stopped
 @property(nonatomic, assign) bool stopped;
-    
+
+/// Ads only
+
+///Ad Manifest file requested
+@property(nonatomic, assign) bool adManifestRequested;
+
 /// Only used for ads
 @property(nonatomic, assign) bool adInitiated;
+
+/// Ad break started
+@property(nonatomic, assign) bool adBreakStarted;
 
 /// ---------------------------------
 /// @name Public methods