+++ YouboraLib/adapter/YBPlaybackFlags.m
@@ -15,10 +15,11 @@
     self = [super init];
     if (self) {
         [self reset];
+        self.adManifestRequested = false; //We don't want to be able to reset this one
     }
     return self;
 }
-
+//TODO: Rethink this method contemplating somehow the ad break flag
 - (void)reset {
     self.preloading = false;
     self.started = false;
@@ -28,6 +29,7 @@
     self.buffering = false;
     
     self.adInitiated = false;
+    //self.adBreakStarted = false; //This on MUST stay commented since we don't want to reset it with the rest
 }
 
 @end