+++ YouboraLib/adapter/YBPlayerAdapter.h
@@ -30,6 +30,20 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
 };
 
 /**
+ * Enum defining the possible Youbora ad manifest error values. This referes to the reason
+ * why a manifest failed.
+ * @see <[YBPlayerAdapter fireManifestError]>
+ */
+typedef NS_ENUM(NSUInteger, YBAdManifestError) {
+    /** No response **/
+    YBAdManifestErrorNoResponse,
+    /** Empty response **/
+    YBAdManifestEmptyResponse,
+    /** Wrong response **/
+    YBAdManifestWrongResponse
+};
+
+/**
  * Main Adapter class. All specific player adapters should extend this class specifying a player
  * class.
  *
@@ -237,13 +251,6 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
 - (NSString *) getVersion;
 
 /**
- * Override to return current ad position (only for ads)
- * @see <YBAdPosition>
- * @return the current ad position
- */
-- (YBAdPosition) getPosition;
-
-/**
  * Override to return household id
  *
  * @return housohold player id
@@ -277,6 +284,99 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
  * @return current p2p state
  */
 -(NSValue *) getIsP2PEnabled;
+
+/// Only ads
+
+/**
+ * Override to return current ad position (only for ads)
+ * @see <YBAdPosition>
+ * @return the current ad position
+ */
+- (YBAdPosition) getPosition;
+
+/**
+ * Override to return current ad break playing
+ * @return current break playing
+ */
+- (nullable NSNumber *) getAdBreakNumber;
+
+/**
+ * Override to return total breaks will play
+ *
+ * @return total ad breaks that will play
+ */
+- (nullable NSNumber *) getAdGivenBreaks;
+
+/**
+ * Override to return total breaks should play
+ *
+ * @return total ad breaks that should play
+ */
+- (nullable NSNumber *) getAdExpectedBreaks;
+
+/**
+ * Override to return the ad structure requested
+ * (list with number of pre, mid, and post breaks) (only ads)
+ *
+ * @return how many braeaks per position (pre,mid,post) will be played
+ */
+- (nullable NSDictionary *) getAdExpectedPattern;
+
+/**
+ * Override to return a list of playheads of ad breaks begin time (only ads)
+ *
+ * @return current time of playaback when an ad break will play
+ */
+- (nullable NSArray *) getAdBreaksTime;
+
+/**
+ * Override to return the number of ads given for the break (only ads)
+ *
+ * @return number of ads on current break
+ */
+- (nullable NSNumber *) getGivenAds;
+
+/**
+ * Override to return the number of ads requested for the break (only ads)
+ *
+ * @return number of expected ads on current break
+ */
+- (nullable NSNumber *) getExpectedAds;
+
+/**
+ * Override to return how long the ad has been watched
+ *
+ * @return how long the ad has been watched
+ */
+- (nullable NSNumber *) getAdViewedDuration;
+
+/**
+ * Override to return maximum time the ad has been watched without interruptions
+ *
+ * @return maximum time the ad has been watched without interruptions
+ */
+- (nullable NSNumber *) getAdViewability;
+
+/**
+ * Override to return if ad can be skipped
+ *
+ * @return boolean indicating if ad can be skipped
+ */
+- (nullable NSValue *) isSkippable;
+
+/**
+ * Override to return ad creative id
+ *
+ * @return string with ad creative id value
+ */
+- (nullable NSString *) getAdCreativeId;
+
+/**
+ * Override to return ad provider
+ *
+ * @return string with ad provider
+ */
+- (nullable NSString *) getAdProvider;
 
 /// ---------------------------------
 /// @name Flow methods
@@ -503,7 +603,48 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
  */
 - (void) fireFatalErrorWithMessage:(nullable NSString *) msg code:(nullable NSString *) code andMetadata:(nullable NSString *) errorMetadata andException:(nullable NSException *)exception;
 
+//Ads only
 
+/**
+ * Sends that a quartile has been reached
+ * @param quartileNumber wich quartile has been reached (must be a number between 1 and 3)
+ */
+- (void) fireQuartile:(int) quartileNumber;
+
+/**
+ * Sends that a manifest has been request successfully
+ */
+- (void) fireAdManifest:(nullable NSDictionary<NSString *, NSString *> *) params;
+
+/**
+ * Sends that a manifest has not been resquested successfully
+ * @see <YBAdManifestError>
+ * @param error error type
+ * @param message error message
+ */
+- (void) fireAdManifestWithError:(YBAdManifestError) error andMessage:(nullable NSString *) message;
+
+/**
+ * Shortcut for <fireBreakStart:> with params = nil.
+ */
+- (void) fireAdBreakStart;
+
+/**
+ * Emits related event and set flags if current status is valid.
+ * @param params Map of key:value pairs to add to the request
+ */
+- (void) fireAdBreakStart:(nullable NSDictionary<NSString *, NSString *> *) params;
+
+/**
+ * Shortcut for <fireBreakStop:> with params = nil.
+ */
+- (void) fireAdBreakStop;
+
+/**
+ * Emits related event and set flags if current status is valid.
+ * @param params Map of key:value pairs to add to the request
+ */
+- (void) fireAdBreakStop:(nullable NSDictionary<NSString *, NSString *> *) params;
 
 /**
  * Adds an adapter delegate that will be called whenever the Adapter
@@ -608,6 +749,13 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
 - (void) youboraAdapterEventError:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
 
 /**
+ * Adapter detected video event
+ * @param params params to add to the request
+ * @param adapter the adapter that is firing the event
+ */
+- (void) youboraAdapterEventVideoEvent:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
+/**
  * Adapter detected ad click
  * @param params params to add to the request
  * @param adapter the adapter that is firing the event
@@ -621,12 +769,36 @@ typedef NS_ENUM(NSUInteger, YBAdPosition) {
  */
 - (void) youboraAdapterEventAllAdsCompleted:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
 
-/**
- * Adapter detected video event
+/** Adapter detected a quartitle has been reached
  * @param params params to add to the request
  * @param adapter the adapter that is firing the event
  */
-- (void) youboraAdapterEventVideoEvent:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+- (void) youboraAdapterEventAdQuartile:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
+/** Adapter deteccted a successful ad manifest request
+ * @param params params to add to the request
+ * @param adapter the adapter that is firing the event
+ */
+- (void) youboraAdapterEventAdManifest:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
+/** Adapter detected a unsuccessful ad manifest request
+ * @param params params to add to the request
+ * @param adapter the adapter that is firing the event
+ */
+- (void) youboraAdapterEventAdManifestError:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
+/** Adapter detected a ad break start
+ * @param params params to add to the request
+ * @param adapter the adapter that is firing the event
+ */
+- (void) youboraAdapterEventAdBreakStart:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
+/** Adapter detected a ad break finished
+ * @param params params to add to the request
+ * @param adapter the adapter that is firing the event
+ */
+- (void) youboraAdapterEventAdBreakStop:(nullable NSDictionary *) params fromAdapter:(YBPlayerAdapter *) adapter;
+
 @end
 
 NS_ASSUME_NONNULL_END