+++ YouboraLib/plugin/YBRequestBuilder.h
@@ -67,6 +67,12 @@ NS_ASSUME_NONNULL_BEGIN;
  */
 - (NSString *) getNewAdNumber;
 
+/**
+ Creates a breakNumber if it does not exist and stores it in lastSent. If it already exists,
+ * it is incremented by 1.
+ */
+- (NSString *) getNewAdBreakNumber;
+
 @end
 
 NS_ASSUME_NONNULL_END;